+++ src/main/java/de/maxhenkel/voicechat/config/ServerConfig.java
@@ -9,6 +9,7 @@ public class ServerConfig {
     public final ConfigBuilder.ConfigEntry<String> voiceChatBindAddress;
     public final ConfigBuilder.ConfigEntry<Double> voiceChatDistance;
     public final ConfigBuilder.ConfigEntry<Double> voiceChatFadeDistance;
+    public final ConfigBuilder.ConfigEntry<Double> crouchDistanceMultiplier;
     public final ConfigBuilder.ConfigEntry<Enum<Codec>> voiceChatCodec;
     public final ConfigBuilder.ConfigEntry<Integer> voiceChatMtuSize;
     public final ConfigBuilder.ConfigEntry<Integer> keepAlive;
@@ -23,6 +24,7 @@ public class ServerConfig {
         voiceChatBindAddress = builder.stringEntry("bind_address", "0.0.0.0");
         voiceChatDistance = builder.doubleEntry("voice_distance", 32D, 1D, 1_000_000D);
         voiceChatFadeDistance = builder.doubleEntry("voice_fade_distance", 16D, 1D, 1_000_000D);
+        crouchDistanceMultiplier = builder.doubleEntry("crouch_distance_multiplier", 1D, 0.01D, 1D);
         voiceChatCodec = builder.enumEntry("codec", Codec.VOIP);
         voiceChatMtuSize = builder.integerEntry("mtu_size", 1024, 256, 10000);
         keepAlive = builder.integerEntry("keep_alive", 1000, 1000, Integer.MAX_VALUE);